requirements.txt
fastapi
uvicorn
groq
python-dotenv
pydantic
beautifulsoup4
lxml
httpx
